Сделал рефакторинг, перенес некоторые общие функции в файл utils.js
