+++ src/mock/movie.js
@@ -1,28 +1,5 @@
 import dayjs from 'dayjs';
-
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-const getRandomFloat = (minValue, maxValue, precision) => {
-  if (minValue < 0 || maxValue < 0) {
-    throw new Error('Min or Max value must be greater or equals to 0');
-  }
-  if (maxValue <= minValue) {
-    throw new Error('Min should be less than max');
-  }
-  return parseFloat(Math.min(minValue + (Math.random() * (maxValue - minValue)),maxValue).toFixed(precision));
-};
-
-const getRandomArrayElement = (array) => array[getRandomInteger(0, array.length - 1)];
-
-const getRandomArray = ([...source], maxLength) => Array.from(
-  { length: Math.min(source.length, 1 + Math.random() * maxLength | 0) },
-  () => source.splice(Math.random() * source.length | 0, 1)[0],
-);
+import {getRandomArray, getRandomArrayElement, getRandomInteger, getRandomFloat} from '../utils';
 
 const comments = [
   {
@@ -40,6 +17,8 @@ const comments = [
     emoji: 'angry.png'
   }
 ];
+
+const generateComments = () => [[comments[0].id], [comments[1].id]];
 
 const generateDescription = () => {
   const descriptions = [
@@ -81,7 +60,7 @@ const generateTitle = () => {
   return getRandomArrayElement(titles);
 };
 
-const generateGenre = () => {
+const generateGenres = () => {
   const genres = [
     'Drama',
     'Film-Noir',
@@ -157,8 +136,6 @@ const generateActors = () => {
   return getRandomArray(actors, getRandomInteger(0, actors.length - 1));
 };
 
-const generateComments = () => [[comments[0].id], [comments[1].id]];
-
 const generateMovie = () => ({
   poster: generatePosters(),
   title: generateTitle(),
@@ -170,7 +147,7 @@ const generateMovie = () => ({
   releaseDate: generateDate(),
   releaseCountry: generateCountry(),
   duration: generateDuration(),
-  genre: generateGenre(),
+  genres: generateGenres(),
   description: generateDescription(),
   ageRating: generateAgeRating(),
   comments: generateComments()
